requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.5
opencv-python>=4.5.3
Pillow>=8.3.1
tqdm>=4.62.0
matplotlib>=3.4.3
pyyaml>=5.4.1
scipy>=1.7.0
albumentations>=1.0.3
tensorboard>=2.7.0
scikit-learn>=0.24.2